Moved the rendering of the frozen status into a separate class
